Suppressing rank deficient warning during spline weight inversion.
Was getting a rank deficient error when inverting spline weights as part of fastBspline.m. We know the matrix is rank deficient I think, which is part of why we use A\b to do the inverse (least squares? Or maybe minimum weights). So I am just supressing the warning... Kind of risky to do this since it might be important elsewhere.