Fixed problem where no models were considered good if too few chains were used. c1_PROCESS_RESULTS.
In  c1_PROCESS_RESULTS, there was a glitch(?) where if less than 3 chains were used, no chains were determined to be good (sometimes?). Not good for debugging. Jon's version had a fix for this, which I copied.

Hopefully this fixed the problem where parallel code could not work for multiple chains.