Progress sending files to ERI computers automatically
Now on either mac or ERI computers, I should be able to run transfer./compileTest_ALL.bash and it will go to propmat, cadmineos and do the compilations and tests. There might be a transfer error occasionally, as PropMat didn't compile the first time I did this and I had to go to the PropMat folder and run ./sendERI.bash from there to get files over properly. Not sure why. I will continue to figure it out...

First, should be able to go to the THBI transfer folder and run sendERI_ALL.bash to send everything to the proper location on ERI computers.