paths variable loading style changes, fixed problem with running inversion on wrong station, H/V ratio progress.
Changes way "paths" variable is loaded to getPaths(). Works from any folder as long as getPaths function is within the Matlab path.

In RUN_ALL_STAS.m I fixed problem where it was not running on the station which I meant for it to run on.

A few changes making progress toward using H/V ratio, but it is not working yet.