Transfer scripts more generic; progress toward parallel run of many stations.
Transfer scripts more generic in terms of which computer they send to.

Make batch folder, which has code to execute several stations in parallel. However, a major problem is that each parallel task will be executing from the ENAM folder, with no way of distinguishing which files belong to which station. Need to next add network.station to each file.