Got entire inversion to run on RAM successfully.
Still needs work in terms of getting the ram folder open on any computer, saving the inversions state to a hard drive every X iterations (I think this is already done with save_inversion anyway... just do it there), and some cleaning up.

Current speed for one chain (with 16 cores and chains in parallel) is about 150 s for 300 iterations. This is basically the same speed as if there is just one chain.