Moved master_par parfor loop content to function + a correction to when surface wave kernels are reset
Long time ago, I changed the surface wave kernel reset mightReset, so that after burnin or maxnkiter got high, we would reset kernels only if the model was accepted. But this isn't the point. Those should execute only if we didn't both accept the last model and make K for it already. Part of the point of this block is resetting the kernels with the OLD model. Mostly see new line 426.