Test: Sp receiver functions, replace parent pulse with a simple Gaussian.
When sediment is present, the parent pulse for Sp receiver functions can be obscured by reverberation. Since we are constructing synthetic receiver functions with a Gaussian pulse, the P component is already a receiver function... so rather than risking deconvolving the parent along with several sediment reverberations, which messes up the cross-convolution error calculation for receiver functions, I'm just introducing a trivial Gaussian parent. The point of having the parent present at all is to make sure the timing in the receiver functions is correct and that the frequency content is compatable between the synthetic and observed receiver functions. The Gaussian pulse accomplishes both of those tasks, more or less.

VERY IMPORTANT propmat has a hard baked factor of Gaussian std = 1/6 * desired pulse width. So I hard coded that into c5 and b3. This will cause a mistake one day in the future, so I made that code as explicit as possible.