HK stack was not getting average vs, vp, rho, xi properly.
There was some systematic offset with the HK stacks. It seems fixed.

It was doing a simple mean of parameters from all depths. This was a problem for the final modelling especially, since tehre was such fine spacing for the sediment and near the Moho. I am now doing an integral style average.