TEMPORARY BUT IMPORTANT Degrees of freedom for Sp ccp totally changes results.
I'm now doing what we were doing for the Sp receiver functions (not ccp) to estimate degrees of freedom. I get about 3x as many degrees of freedom (something like from 17 to 50). The change makes it so we can fit (noisy) synthetic data, and get all the MLDs and stuff I put in the synthetic model. Before, I was not recovering the MLDs.