Testing different approaches for getting degrees of freedom for real CCP receiver functions.
dofTHBI_SW_low.: gets very low degree of freedom for surface wave data sets. It's just a test. I expect each surface wave period to be rather dependent on the others, so the old assumption that there is one degree of freedom for each period might be inconsistent with our more realistic (and pessimistic) estimation of degrees of freedom for receiver functions.