Inversion now runs through without error, in parallel, and over final forward model. Ram fix.
Fix problem with ram folder filling up when using parallel loop on chains

Needed to delete the Mineos files after getting kernels. I also increased the amount of ram available in the ram folder, and this scales depending on the number of processors.

Was getting random and confusing errors, at different stages of MCMC on different chains, and ram filled up.